ffi module cleanup: listobject.h to memoryobject.h
