+++ src/ffi/cpython/mod.rs
@@ -9,6 +9,9 @@ pub mod dictobject;
 // skipped fileobject.h
 pub mod frameobject;
 // skipped import.h
+// skipped initconfig.h
+// skipped interpreteridobject.h
+pub mod listobject;
 
 pub use self::abstract_::*;
 #[cfg(not(PyPy))]
@@ -18,3 +21,4 @@ pub use self::code::*;
 #[cfg(not(PyPy))]
 pub use self::dictobject::*;
 pub use self::frameobject::*;
+pub use self::listobject::*;
