+++ src/ffi/listobject.rs
@@ -2,14 +2,6 @@ use crate::ffi::object::*;
 use crate::ffi::pyport::Py_ssize_t;
 use std::os::raw::c_int;
 
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub struct PyListObject {
-    pub ob_base: PyVarObject,
-    pub ob_item: *mut *mut PyObject,
-    pub allocated: Py_ssize_t,
-}
-
 #[cfg_attr(windows, link(name = "pythonXY"))]
 extern "C" {
     #[cfg_attr(PyPy, link_name = "PyPyList_Type")]
@@ -26,26 +18,6 @@ pub unsafe fn PyList_Check(op: *mut PyObject) -> c_int {
 #[inline]
 pub unsafe fn PyList_CheckExact(op: *mut PyObject) -> c_int {
     (Py_TYPE(op) == &mut PyList_Type) as c_int
-}
-
-/// Macro, trading safety for speed
-#[cfg(not(Py_LIMITED_API))]
-#[inline]
-pub unsafe fn PyList_GET_ITEM(op: *mut PyObject, i: Py_ssize_t) -> *mut PyObject {
-    *(*(op as *mut PyListObject)).ob_item.offset(i as isize)
-}
-
-#[cfg(not(Py_LIMITED_API))]
-#[inline]
-pub unsafe fn PyList_GET_SIZE(op: *mut PyObject) -> Py_ssize_t {
-    Py_SIZE(op)
-}
-
-/// Macro, *only* to be used to fill in brand new lists
-#[cfg(not(Py_LIMITED_API))]
-#[inline]
-pub unsafe fn PyList_SET_ITEM(op: *mut PyObject, i: Py_ssize_t, v: *mut PyObject) {
-    *(*(op as *mut PyListObject)).ob_item.offset(i as isize) = v;
 }
 
 extern "C" {
