+++ src/ffi/memoryobject.rs
@@ -2,6 +2,8 @@ use crate::ffi::object::*;
 use crate::ffi::pyport::Py_ssize_t;
 use std::os::raw::{c_char, c_int};
 
+// skipped non-limited _PyManagedBuffer_Type
+
 #[cfg_attr(windows, link(name = "pythonXY"))]
 extern "C" {
     #[cfg_attr(PyPy, link_name = "PyPyMemoryView_Type")]
@@ -13,6 +15,9 @@ pub unsafe fn PyMemoryView_Check(op: *mut PyObject) -> c_int {
     (Py_TYPE(op) == &mut PyMemoryView_Type) as c_int
 }
 
+// skipped non-limited PyMemoryView_GET_BUFFER
+// skipped non-limited PyMemeryView_GET_BASE
+
 extern "C" {
     #[cfg_attr(PyPy, link_name = "PyPyMemoryView_FromObject")]
     pub fn PyMemoryView_FromObject(base: *mut PyObject) -> *mut PyObject;
@@ -22,9 +27,15 @@ extern "C" {
         size: Py_ssize_t,
         flags: c_int,
     ) -> *mut PyObject;
+    // skipped non-limited PyMemoryView_FromBuffer
     pub fn PyMemoryView_GetContiguous(
         base: *mut PyObject,
         buffertype: c_int,
         order: c_char,
     ) -> *mut PyObject;
 }
+
+// skipped remainder of file with comment:
+/* The structs are declared here so that macros can work, but they shouldn't
+be considered public. Don't access their fields directly, use the macros
+and functions instead! */
